Migrate to PyTorch tokenizers (#781)
## Description

This PR migrates us from tokenisers-cpp to PyTorch tokenisers that are by default bundled
with executorch binaries

### Introduces a breaking change?

- [ ] Yes
- [x] No - User faces no changes 

### Type of change

- [ ] Bug fix (change which fixes an issue)
- [ ] New feature (change which adds functionality)
- [ ] Documentation update (improves or adds clarity to existing
documentation)
- [x] Other (chores, tests, code style improvements etc.) 

### Tested on

- [x] iOS
- [x] Android

### Testing instructions

This changes need to be tested manually. Try running all our apps that
consume tokenizers and see whether the output is ok.

### Related issues

<!-- Link related issues here using #issue-number -->

### Checklist

- [x] I have performed a self-review of my code
- [x] I have commented my code, particularly in hard-to-understand areas
- [ ] I have updated the documentation accordingly
- [ ] My changes generate no new warnings

+++ packages/react-native-executorch/common/rnexecutorch/models/speech_to_text/SpeechToText.h
@@ -22,7 +22,7 @@ public:
   encode(std::span<float> waveform) const;
   [[nodiscard(
       "Registered non-void function")]] std::shared_ptr<OwningArrayBuffer>
-  decode(std::span<int32_t> tokens, std::span<float> encoderOutput) const;
+  decode(std::span<uint64_t> tokens, std::span<float> encoderOutput) const;
   [[nodiscard("Registered non-void function")]] std::vector<char>
   transcribe(std::span<float> waveform, std::string languageOption) const;
